Character tree refactoring
Player character should be fully in place after refactoring now.
