+++ Content/Art/MetaHumans/Ettore/Previews/BP_Ettore_PreviewMeshCollection.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9709b79149c800fc998172ddf9b9c2f77974ec5101b3831e0ff240fb687d04e3
-size 3184
+oid sha256:d9d31524d0504ace081c2eadb467c2f4cc049b0d527f1bf8a5613e7331071fcb
+size 3177
