+++ Content/Blueprints/Characters/ABP_EclipseRaptureCharacter.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1f5e55da1b9896c82bb717a28de53761312e31f0e7293f2301e42917e553f578
+size 1775889
