--- Content/Blueprints/Characters/Player/ABP_EclipseRaptureCharacter.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b49b248c6d144d8ccc45a22eaf2e75c6b4b9bde2640f223c70eda7e0ed4e6136
-size 1768614
