--- Content/Blueprints/Characters/Player/BP_EclipseRaptureCharacter.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9f714a44971b45e3e9f485f63a9800d55d9fecf0e9069f2d319c1089ca07d934
-size 112883
