+++ Content/Blueprints/Characters/Player/BP_PlayerMain.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5978b2ef6d5e32bc310a018514ea8e91afc00b8e421fc78366eff22dacf848e1
-size 2718656
+oid sha256:bc49d0454cee55e0d1b55de30338ee2ff99b175cf6db21732308c2c36a665e01
+size 2797932
