+++ Content/Blueprints/Items/Interactables/OneTimeInteractions/BP_PlayerFlashlight.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7dcb656c8c8e1e06114148068bfe8bd685186812b8baa38db14219ff128c2940
-size 58086
+oid sha256:c507a9e691ca5e17c95157b53da356259b64e7c853fe9fb2d10d5188bfd189fc
+size 62091
