+++ Content/Blueprints/Items/Weapons/Weapons/BP_WeaponBase.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4245c03b19e72434f6dd098bf0be753899065feb0eb39106bfda5afc43051d94
-size 296146
+oid sha256:e86ac09c02e7460977ae68c0e437babcb1f4e0544383ecc21e5a660d0df2f948
+size 294283
