+++ Content/Blueprints/UI/Player/BPW_Inventory.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6f1b1aecc8f0ec5d8d4f5108cd1199aa2b5adc3860449a3a72f0d5a4e5bd8982
-size 147282
+oid sha256:c08e01803632de9f0977a5b6b4b8f6af5b542a5112ee2bb389a977e230bbdaa5
+size 148180
