+++ Content/Blueprints/UI/Player/BPW_InventoryItem.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c3b891fd561921f31b81bbd02b0913f831e610616223596afee87d12324f1503
-size 148614
+oid sha256:01da02935acef8dd34b081f8164e7c55df2256235aa847aeaba3253a59cd1c7b
+size 151633
