+++ Content/Blueprints/UI/Player/BPW_PlayerMain.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:213f12f867ac67f3fbdef40f31bfa1901f869f3bf6bc6c90c7d83c91c1386119
-size 262268
+oid sha256:bb431a1fadeb81da90dc97ffe7e2e87625ec161d050667a83bf8bcb490cf1b86
+size 262970
