+++ Source/Eclipse_Rapture/Private/Items/Item.cpp
@@ -4,6 +4,7 @@
 #include "Items/Item.h"
 #include "Components/SphereComponent.h"
 #include "Interfaces/IPhysicsComponent.h"
+#include "Character/PlayerMain.h"
 
 AItem::AItem()
 {
@@ -59,7 +60,7 @@ void AItem::Use(AEclipseRaptureCharacter* Character)
 
 void AItem::OnSphereOverlap(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult)
 {
-	AEclipseRaptureCharacter* Character = Cast<AEclipseRaptureCharacter>(OtherActor);
+	APlayerMain* Character = Cast<APlayerMain>(OtherActor);
 	if (Character)
 	{
 		Character->SetCurrentlyOverlappingItem(this);
@@ -68,7 +69,7 @@ void AItem::OnSphereOverlap(UPrimitiveComponent* OverlappedComponent, AActor* Ot
 
 void AItem::OnSphereEndOverlap(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex)
 {
-	AEclipseRaptureCharacter* Character = Cast<AEclipseRaptureCharacter>(OtherActor);
+	APlayerMain* Character = Cast<APlayerMain>(OtherActor);
 	if (Character)
 	{
 		Character->SetCurrentlyOverlappingItem(nullptr);
